requirements.txt
fastapi
uvicorn
requests
sqlalchemy
google-cloud-translate
jinja2  